gemini_quiz_app/requirements.txt
Flask>=3.0.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
gunicorn